requirements.txt
aiofiles==24.1.0                 # File handling in async frameworks
anyio==3.7.1                     # Compatibility layer for multiple async I/O frameworks
boto3==1.26.105                  # AWS SDK for Python to interact with S3 and other services
botocore==1.29.165              # Core functionality for boto3
certifi==2025.7.14              # Provides Mozilla’s root certificates for secure HTTPS
charset-normalizer==3.4.2       # Used for encoding detection in requests
click==8.2.1                    # Command-line interface creation toolkit
colorama==0.4.6                 # Cross-platform colored terminal text
contourpy==1.3.2                # Backend for contour plots in matplotlib
cycler==0.12.1                  # Composable cycles used by matplotlib
Cython==3.1.2                   # Translates Python code to C for performance boost
dill==0.3.5.1                   # Advanced serialization library, better than pickle
dnspython==2.2.1                # DNS toolkit for Python (used with pymongo sometimes)
evidently==0.3.3                # Monitoring ML models and data quality in production
exceptiongroup==1.3.0           # Provides a way to group multiple exceptions (used in async)
Flask==2.2.3                    # Lightweight web framework for Python
fonttools==4.59.0               # Manipulate font files (used by matplotlib)
h11==0.16.0                     # HTTP/1.1 protocol used by async frameworks like uvicorn
httptools==0.5.0                # Fast HTTP parser (used in uvicorn)
idna==3.10                      # Internationalized domain name support for URLs
imbalanced-learn==0.13.0        # Handle imbalanced datasets in ML
imblearn==0.0                   # Meta-package for imbalanced-learn
itsdangerous==2.2.0             # Used by Flask for secure session tokens
Jinja2==3.1.2                   # Template engine used by Flask
jmespath==1.0.1                 # JSON query language used by boto3
joblib==1.5.1                   # Lightweight pipelining and parallelization in Python
kiwisolver==1.4.8               # Solver for layout in matplotlib
kneed==0.8.2                    # Detect the "knee" or "elbow" point in graphs (useful for clustering)
llvmlite==0.44.0                # Required for JIT compilation in numba
MarkupSafe==3.0.2               # Used by Jinja2 for safe string handling
matplotlib==3.7.1               # Data visualization and plotting
mypy-boto3-s3==1.24.76          # Type hints for boto3 S3 (helpful during development)
narwhals==1.48.0                # Used internally by evidently for profiling and comparison
neuro-mf==0.0.5                 # Matrix factorization used in recommendation systems
nltk==3.9.1                     # Natural Language Toolkit for NLP tasks
numba==0.61.2                   # JIT compiler for numerical functions
numpy>=1.21,<2.0                  # Fundamental numerical computing library
packaging==25.0                 # Helps in version comparison and package metadata
pandas==2.3.1                   # Data manipulation and analysis library
patsy==1.0.1                    # Used with statsmodels for describing statistical models
pillow==11.3.0                  # Python Imaging Library (PIL fork), used in visual tasks
pip-chill==1.0.1                # Generates minimal requirements.txt with only used packages
plotly==6.2.0                   # Interactive plotting and dashboarding
pydantic==1.10.22               # Data validation using Python type hints
pymongo==4.2.0                  # MongoDB driver for Python
pynndescent==0.5.13             # Fast nearest neighbor descent algorithm (used by umap)
pyparsing==3.2.3                # Used for syntactic parsing in packages like matplotlib
python-dateutil==2.9.0.post0    # Enhanced datetime support
python-dotenv==0.21.0           # Loads environment variables from .env files
python-multipart==0.0.6         # Parses multipart/form-data (file uploads)
pytz==2025.2                    # World timezone definitions
PyYAML==6.0.2                   # YAML parser and emitter for Python
regex==2024.11.6                # Extended regular expression support
requests==2.32.4                # Simplified HTTP requests library
s3transfer==0.6.2               # Manages S3 file transfers (used by boto3)
scikit-learn==1.6.1             # Machine learning library
scipy==1.15.3                   # Scientific computing and numerical analysis
seaborn==0.12.2                 # Statistical data visualization built on matplotlib
six==1.17.0                     # Python 2 and 3 compatibility library
sklearn-compat==0.1.3           # Compatibility layer between sklearn and other libraries
sniffio==1.3.1                  # Detect which async library is in use
statsmodels==0.14.5             # Statistical modeling and hypothesis testing
threadpoolctl==3.6.0            # Controls native libraries for threading
tqdm==4.67.1                    # Progress bar for loops and processes
types-s3transfer==0.6.0.post4   # Type hints for s3transfer (for development)
typing_extensions==4.14.1       # Backport of new typing features for older Python versions
tzdata==2025.2                  # Time zone database for date/time handling
umap-learn==0.5.9.post2         # Dimensionality reduction using UMAP
urllib3==1.26.20                # HTTP client used by requests
uvicorn==0.18.3                 # ASGI server for serving FastAPI/Flask async apps
watchfiles==0.17.0              # Watch filesystem for changes (useful in dev servers)
websockets==10.3                # WebSocket protocol support
Werkzeug==3.1.3                 # WSGI toolkit used by Flask
wincertstore==0.2               # Windows certificate store access (for HTTPS)
xgboost==1.6.2                  # Gradient boosting library for classification and regression
ipykernel